+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - Update healthcheck routes from `/v1/healthcheck` to `/healthz`
+
 ## [0.0.18] - 2025-11-14
 
 ### Added