+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.19] - 2025-11-16
+
 ### Added
  - Logging middleware in both control and streaming servers (#1032)
 
@@ -81,7 +83,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.10] - 2025-02-05
 
-[unreleased]: https://github.com/skiplabs/skip/compare/v0.0.18...HEAD
+[unreleased]: https://github.com/skiplabs/skip/compare/v0.0.19...HEAD
+[0.0.19]: https://github.com/skiplabs/skip/compare/v0.0.18...v0.0.19
 [0.0.18]: https://github.com/skiplabs/skip/compare/v0.0.17...v0.0.18
 [0.0.17]: https://github.com/skiplabs/skip/compare/v0.0.16...v0.0.17
 [0.0.16]: https://github.com/skiplabs/skip/compare/v0.0.15...v0.0.16